add tests and clean routes in order, sales and admin

+++ app/routes/admin.py
@@ -4,7 +4,6 @@ from app.models import Operation, db
 from app.schemas import BookSchema, OperationCancelSchema, OperationSchema
 from app.utils.decorators import role_required
 from app.utils.helpers import cancel_operation
-from datetime import date
 
 admin_bp = Blueprint("admin", __name__, url_prefix="/api/admin")
 
@@ -24,37 +23,6 @@ def confirm_order(order_id: int):
 
     return jsonify({"msg": "Order confirmed"}), 200
 
-@admin_bp.route("/operations", methods=["DELETE"])
-@jwt_required()
-@role_required("admin")
-def cancel_op_admin():
-    data = OperationCancelSchema().load(request.json)
-    op = cancel_operation(data["customer_id"], data["op_date"], data["op_type"], is_admin=True)
-    return OperationSchema().dump(op), 200
-
-
-    # op = Operation.query.get(operation_id)
-    # if op is None:
-    #     return jsonify({ "msg": "Not found" }), 404
-    
-    # schema = OperationCancelSchema()
-    # confirmation_data = request.get_json()
-    # errors = schema.validate(confirmation_data)
-    # if errors:
-    #     return jsonify(errors)
-    
-    # op_date: str = confirmation_data["delete_date"]
-    # print(op_date)
-    # [year, mm, dd] = map(int, op_date.split("-"))
-    
-    # if (op.date == date(year=year, month=mm, day=dd) and op.customer_id == confirmation_data["delete_customer_id"]):
-    #     db.session.delete(op)
-    #     db.session.commit()
-    # else:
-    #     return { "msg": "No match, either date or customer_id" }, 200
-
-    # schema = OperationSchema()
-    # return schema.dump(op), 204
 
 @admin_bp.route("/operations/<int:operation_id>", methods=["DELETE"])
 @jwt_required()
@@ -63,9 +31,16 @@ def delete_operation(operation_id):
     op = Operation.query.get(operation_id)
     if not op:
         return jsonify({"msg": "Not found"}), 404
+    
+    if op.op_type == "delivered":
+        op.op_type = "cancelled"
+        db.session.commit()
+        return OperationSchema().dump(op), 200
+    
     db.session.delete(op)
     db.session.commit()
-    return OperationSchema().dump(op), 204
+    return "", 204
+
 
 @admin_bp.route("/operations", methods=["GET"])
 @jwt_required()
@@ -73,6 +48,7 @@ def delete_operation(operation_id):
 def all_operations():
     all_op = Operation.query.all()
     return OperationSchema(many=True).dump(all_op)
+
 
 
 @admin_bp.route("/books", methods=["POST"])
